remove unnecessary header file
